fix : chart view와 로직 관심사 분리
